vm: consolidate precompile routing into precompiles.go
Move fork-keyed precompile maps, ActivePrecompiles, ActivePrecompiledContracts,
GetPrecompiledContractMap, and getPrecompiledContractForVersion from contracts.go
and evm.go into a new blockchain/vm/precompiles.go. Add a design note explaining
the three distinct views (range-check, access-list, execution). Add a cross-reference
comment to common/types.go pointing to precompiles.go for the full routing logic.
