+++ blockchain/vm/contracts.go
@@ -80,198 +80,6 @@ type PrecompiledContract interface {
 	Name() string
 }
 
-// PrecompiledContractsByzantium contains the default set of pre-compiled Kaia
-// contracts based on Ethereum Byzantium.
-var PrecompiledContractsByzantium = map[common.Address]PrecompiledContract{
-	common.BytesToAddress([]byte{1}):  &ecrecover{},
-	common.BytesToAddress([]byte{2}):  &sha256hash{},
-	common.BytesToAddress([]byte{3}):  &ripemd160hash{},
-	common.BytesToAddress([]byte{4}):  &dataCopy{},
-	common.BytesToAddress([]byte{5}):  &bigModExp{eip2565: false, eip7823: false, eip7883: false},
-	common.BytesToAddress([]byte{6}):  &bn256AddByzantium{},
-	common.BytesToAddress([]byte{7}):  &bn256ScalarMulByzantium{},
-	common.BytesToAddress([]byte{8}):  &bn256PairingByzantium{},
-	common.BytesToAddress([]byte{9}):  &vmLog{},
-	common.BytesToAddress([]byte{10}): &feePayer{},
-	common.BytesToAddress([]byte{11}): &validateSender{},
-}
-
-// DO NOT USE 0x3FD, 0x3FE, 0x3FF ADDRESSES BEFORE ISTANBUL CHANGE ACTIVATED.
-
-// PrecompiledContractsIstanbul contains the default set of pre-compiled Kaia
-// contracts based on Ethereum Istanbul.
-var PrecompiledContractsIstanbul = map[common.Address]PrecompiledContract{
-	common.BytesToAddress([]byte{1}):      &ecrecover{},
-	common.BytesToAddress([]byte{2}):      &sha256hash{},
-	common.BytesToAddress([]byte{3}):      &ripemd160hash{},
-	common.BytesToAddress([]byte{4}):      &dataCopy{},
-	common.BytesToAddress([]byte{5}):      &bigModExp{eip2565: false, eip7823: false, eip7883: false},
-	common.BytesToAddress([]byte{6}):      &bn256AddIstanbul{},
-	common.BytesToAddress([]byte{7}):      &bn256ScalarMulIstanbul{},
-	common.BytesToAddress([]byte{8}):      &bn256PairingIstanbul{},
-	common.BytesToAddress([]byte{9}):      &blake2F{},
-	common.BytesToAddress([]byte{3, 253}): &vmLog{},
-	common.BytesToAddress([]byte{3, 254}): &feePayer{},
-	common.BytesToAddress([]byte{3, 255}): &validateSender{},
-}
-
-// PrecompiledContractsKore contains the default set of pre-compiled Kaia
-// contracts based on Ethereum Berlin.
-var PrecompiledContractsKore = map[common.Address]PrecompiledContract{
-	common.BytesToAddress([]byte{1}):      &ecrecover{},
-	common.BytesToAddress([]byte{2}):      &sha256hash{},
-	common.BytesToAddress([]byte{3}):      &ripemd160hash{},
-	common.BytesToAddress([]byte{4}):      &dataCopy{},
-	common.BytesToAddress([]byte{5}):      &bigModExp{eip2565: true, eip7823: false, eip7883: false},
-	common.BytesToAddress([]byte{6}):      &bn256AddIstanbul{},
-	common.BytesToAddress([]byte{7}):      &bn256ScalarMulIstanbul{},
-	common.BytesToAddress([]byte{8}):      &bn256PairingIstanbul{},
-	common.BytesToAddress([]byte{9}):      &blake2F{},
-	common.BytesToAddress([]byte{3, 253}): &vmLog{},
-	common.BytesToAddress([]byte{3, 254}): &feePayer{},
-	common.BytesToAddress([]byte{3, 255}): &validateSender{},
-}
-
-// PrecompiledContractsCancun contains the default set of pre-compiled Kaia
-// contracts based on Ethereum Cancun.
-var PrecompiledContractsCancun = map[common.Address]PrecompiledContract{
-	common.BytesToAddress([]byte{1}):      &ecrecover{},
-	common.BytesToAddress([]byte{2}):      &sha256hash{},
-	common.BytesToAddress([]byte{3}):      &ripemd160hash{},
-	common.BytesToAddress([]byte{4}):      &dataCopy{},
-	common.BytesToAddress([]byte{5}):      &bigModExp{eip2565: true, eip7823: false, eip7883: false},
-	common.BytesToAddress([]byte{6}):      &bn256AddIstanbul{},
-	common.BytesToAddress([]byte{7}):      &bn256ScalarMulIstanbul{},
-	common.BytesToAddress([]byte{8}):      &bn256PairingIstanbul{},
-	common.BytesToAddress([]byte{9}):      &blake2F{},
-	common.BytesToAddress([]byte{0x0a}):   &kzgPointEvaluation{},
-	common.BytesToAddress([]byte{3, 253}): &vmLog{},
-	common.BytesToAddress([]byte{3, 254}): &feePayer{},
-	common.BytesToAddress([]byte{3, 255}): &validateSender{},
-}
-
-// PrecompiledContractsPrague contains the set of pre-compiled Ethereum
-// contracts used in the Prague release.
-var PrecompiledContractsPrague = map[common.Address]PrecompiledContract{
-	common.BytesToAddress([]byte{1}):      &ecrecover{},
-	common.BytesToAddress([]byte{2}):      &sha256hash{},
-	common.BytesToAddress([]byte{3}):      &ripemd160hash{},
-	common.BytesToAddress([]byte{4}):      &dataCopy{},
-	common.BytesToAddress([]byte{5}):      &bigModExp{eip2565: true, eip7823: false, eip7883: false},
-	common.BytesToAddress([]byte{6}):      &bn256AddIstanbul{},
-	common.BytesToAddress([]byte{7}):      &bn256ScalarMulIstanbul{},
-	common.BytesToAddress([]byte{8}):      &bn256PairingIstanbul{},
-	common.BytesToAddress([]byte{9}):      &blake2F{},
-	common.BytesToAddress([]byte{0x0a}):   &kzgPointEvaluation{},
-	common.BytesToAddress([]byte{0x0b}):   &bls12381G1Add{},
-	common.BytesToAddress([]byte{0x0c}):   &bls12381G1MultiExp{},
-	common.BytesToAddress([]byte{0x0d}):   &bls12381G2Add{},
-	common.BytesToAddress([]byte{0x0e}):   &bls12381G2MultiExp{},
-	common.BytesToAddress([]byte{0x0f}):   &bls12381Pairing{},
-	common.BytesToAddress([]byte{0x10}):   &bls12381MapG1{},
-	common.BytesToAddress([]byte{0x11}):   &bls12381MapG2{},
-	common.BytesToAddress([]byte{3, 253}): &vmLog{},
-	common.BytesToAddress([]byte{3, 254}): &feePayer{},
-	common.BytesToAddress([]byte{3, 255}): &validateSender{},
-}
-
-// PrecompiledContractsOsaka contains the set of pre-compiled Ethereum
-// contracts used in the Osaka release.
-var PrecompiledContractsOsaka = map[common.Address]PrecompiledContract{
-	common.BytesToAddress([]byte{1}):         &ecrecover{},
-	common.BytesToAddress([]byte{2}):         &sha256hash{},
-	common.BytesToAddress([]byte{3}):         &ripemd160hash{},
-	common.BytesToAddress([]byte{4}):         &dataCopy{},
-	common.BytesToAddress([]byte{5}):         &bigModExp{eip2565: true, eip7823: true, eip7883: true},
-	common.BytesToAddress([]byte{6}):         &bn256AddIstanbul{},
-	common.BytesToAddress([]byte{7}):         &bn256ScalarMulIstanbul{},
-	common.BytesToAddress([]byte{8}):         &bn256PairingIstanbul{},
-	common.BytesToAddress([]byte{9}):         &blake2F{},
-	common.BytesToAddress([]byte{0x0a}):      &kzgPointEvaluation{},
-	common.BytesToAddress([]byte{0x0b}):      &bls12381G1Add{},
-	common.BytesToAddress([]byte{0x0c}):      &bls12381G1MultiExp{},
-	common.BytesToAddress([]byte{0x0d}):      &bls12381G2Add{},
-	common.BytesToAddress([]byte{0x0e}):      &bls12381G2MultiExp{},
-	common.BytesToAddress([]byte{0x0f}):      &bls12381Pairing{},
-	common.BytesToAddress([]byte{0x10}):      &bls12381MapG1{},
-	common.BytesToAddress([]byte{0x11}):      &bls12381MapG2{},
-	common.BytesToAddress([]byte{0x1, 0x00}): &p256Verify{},
-	common.BytesToAddress([]byte{3, 253}):    &vmLog{},
-	common.BytesToAddress([]byte{3, 254}):    &feePayer{},
-	common.BytesToAddress([]byte{3, 255}):    &validateSender{},
-}
-
-// PrecompiledContractsP256Verify contains the precompiled Ethereum
-// contract specified in EIP-7212. This is exported for testing purposes.
-var PrecompiledContractsP256Verify = map[common.Address]PrecompiledContract{
-	common.BytesToAddress([]byte{0x1, 0x00}): &p256Verify{},
-}
-
-var (
-	PrecompiledAddressOsaka       []common.Address
-	PrecompiledAddressPrague      []common.Address
-	PrecompiledAddressCancun      []common.Address
-	PrecompiledAddressIstanbul    []common.Address
-	PrecompiledAddressesByzantium []common.Address
-)
-
-func init() {
-	for k := range PrecompiledContractsByzantium {
-		PrecompiledAddressesByzantium = append(PrecompiledAddressesByzantium, k)
-	}
-	for k := range PrecompiledContractsIstanbul {
-		PrecompiledAddressIstanbul = append(PrecompiledAddressIstanbul, k)
-	}
-	for k := range PrecompiledContractsCancun {
-		PrecompiledAddressCancun = append(PrecompiledAddressCancun, k)
-	}
-	for k := range PrecompiledContractsPrague {
-		PrecompiledAddressPrague = append(PrecompiledAddressPrague, k)
-	}
-	for k := range PrecompiledContractsOsaka {
-		PrecompiledAddressOsaka = append(PrecompiledAddressOsaka, k)
-	}
-}
-
-// ActivePrecompiles returns the precompiles enabled with the current configuration.
-// This is a variable so that it can be overridden in tests (e.g. EEST) to exclude
-// the vmversion0 compatibility addresses from the access list.
-var ActivePrecompiles = func(rules params.Rules) []common.Address {
-	var precompiledContractAddrs []common.Address
-	for addr := range ActivePrecompiledContracts(rules) {
-		precompiledContractAddrs = append(precompiledContractAddrs, addr)
-	}
-	// After istanbulCompatible hf, need to support for vmversion0 contracts, too.
-	// VmVersion0 contracts are deployed before istanbulCompatible and they use byzantiumCompatible precompiled contracts.
-	// VmVersion0 contracts are the contracts deployed before istanbulCompatible hf.
-	if rules.IsIstanbul {
-		return append(precompiledContractAddrs,
-			[]common.Address{common.BytesToAddress([]byte{10}), common.BytesToAddress([]byte{11})}...)
-	} else {
-		return precompiledContractAddrs
-	}
-}
-
-// ActivePrecompiledContracts returns the precompiled contracts enabled with the current configuration.
-// This function doesn't support for vmversion0 contracts, it only supports for istanbulCompatible hf.
-func ActivePrecompiledContracts(rules params.Rules) map[common.Address]PrecompiledContract {
-	var precompiledContracts map[common.Address]PrecompiledContract
-	switch {
-	case rules.IsOsaka:
-		precompiledContracts = PrecompiledContractsOsaka
-	case rules.IsPrague:
-		precompiledContracts = PrecompiledContractsPrague
-	case rules.IsCancun:
-		precompiledContracts = PrecompiledContractsCancun
-	case rules.IsIstanbul:
-		precompiledContracts = PrecompiledContractsIstanbul
-	default:
-		precompiledContracts = PrecompiledContractsByzantium
-	}
-
-	return precompiledContracts
-}
-
 // RunPrecompiledContract runs and evaluates the output of a precompiled contract.
 func RunPrecompiledContract(p PrecompiledContract, input []byte, contract *Contract, evm *EVM) (ret []byte, computationCost uint64, err error) {
 	gas, computationCost := p.GetRequiredGasAndComputationCost(input)
